refactor: add Demo and Single components to miscellaneous module and update routes
