+++ src/front/modules/miscellaneous/4geeks-defatult-files/Demo.jsx
@@ -1,6 +1,6 @@
 // Import necessary components from react-router-dom and other parts of the application.
 import { Link } from "react-router-dom";
-import useGlobalReducer from "../shared/hooks/useGlobalReducer";  // Custom hook for accessing the global state.
+import useGlobalReducer from "../../../shared/hooks/useGlobalReducer";  // Custom hook for accessing the global state.
 
 export const Demo = () => {
   // Access the global state and dispatch function using the useGlobalReducer hook.
